requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asyncpg==0.31.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
fastapi==0.128.0
filelock==3.20.3
fsspec==2026.1.0
GeoAlchemy2==0.18.1
greenlet==3.3.0
h11==0.16.0
hf-xet==1.2.0
huggingface-hub==0.36.0
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.1
packaging==25.0
pandas==2.3.3
pandas-stubs==2.3.3.251219
psycopg==3.3.2
psycopg-binary==3.3.2
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.21
pytz==2025.2
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.0
setuptools==80.9.0
shapely==2.1.2
six==1.17.0
SQLAlchemy==2.0.45
starlette==0.50.0
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.9.1
tqdm==4.67.1
transformers==4.57.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.40.0
